+++ server/src/embeddedSupport/embeddedSupport.ts
@@ -1,7 +1,7 @@
 import { removeQuotes } from '../utils/strings';
-import { createScanner } from './template/parser/htmlScanner';
+import { createScanner } from '../modes/template/parser/htmlScanner';
 import { TextDocument, Position, Range } from 'vscode-languageserver-types';
-import { TokenType, Scanner } from './template/parser/htmlScanner';
+import { TokenType, Scanner } from '../modes/template/parser/htmlScanner';
 
 export interface LanguageRange extends Range {
   languageId: string;
