+++ server/src/modes/script/serviceHost.ts
@@ -4,7 +4,7 @@ import Uri from 'vscode-uri';
 import { TextDocument } from 'vscode-languageserver-types';
 import * as parseGitIgnore from 'parse-gitignore';
 
-import { LanguageModelCache } from '../languageModelCache';
+import { LanguageModelCache } from '../../embeddedSupport/languageModelCache';
 import { createUpdater, parseVue, isVue } from './preprocess';
 import { getFileFsPath, getFilePath } from '../../utils/paths';
 import * as bridge from './bridge';
